refactor(#92): infinity query hook 디렉토리 분리
